rename 'gerber_doc' module.
